+++ lib/widgets/memo_list_page.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+
+class MemoListPage extends StatefulWidget {
+  const MemoListPage({Key? key}) : super(key: key);
+
+  @override
+  State<MemoListPage> createState() => _MemoListPageState();
+}
+
+class _MemoListPageState extends State<MemoListPage> {
+  void _createMemo() {
+    // TODO: 新規メモ作成ページを表示
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('飲んだお酒のメモ'),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            const Text(
+              'To Do',
+            ),
+          ],
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: _createMemo,
+        tooltip: 'メモを追加します。',
+        child: const Icon(Icons.add),
+      ),
+    );
+  }
+}
